+++ attributes/mattermost.rb
@@ -1,6 +1,6 @@
 default['mattermost']['package'] = {
-  'url' => 'https://releases.mattermost.com/4.9.2/mattermost-4.9.2-linux-amd64.tar.gz',
-  'checksum' => '51cc30502eb1ae84742098db26c22bdfb464849bfdd8b3597d950e389f090839',
+  'url' => 'https://releases.mattermost.com/4.10.1/mattermost-4.10.1-linux-amd64.tar.gz',
+  'checksum' => 'cef8a706d6da1d8756d70d06a9e9444ba078fb107a194ce91ea2e6beae9726f7',
 }
 
 default['mattermost']['config'] = {